==> AmazonBookReview/1 - First analyse.ipynb <==
import pandas as pd
import numpy as np

df=pd.read_csv('data/ratings_Books.csv',usecols=[0,1,2],names=['userId','itemId','rating'])

df.head()

df.to_parquet('data/ratings_Books.gzip',compression='gzip',index=False)

df=pd.read_parquet('data/ratings.gzip')

len(df)

ar=df.to_numpy()

dicoId={}
dicoAr={}
t=len(ar)
idCpt=1
for i in range(t):
    ID=ar[i,0]
    if ID in dicoId :
        dicoAr[dicoId[ID]].append(np.array([[dicoId[ID],ar[i,1],ar[i,2]]]))
    else :
        dicoId[ID]=idCpt
        dicoAr[idCpt]=list()
        dicoAr[idCpt].append(np.array([[dicoId[ID],ar[i,1],ar[i,2]]]))
        idCpt+=1

listAR=list()
for i in dicoAr:
    listAR+=dicoAr[i]

len(listAR)==len(df)

len(df)

listARbis=list()
pas=100000
n=len(listAR)//pas
for i in range(n):
    listARbis.append(np.concatenate(listAR[i*pas:(i+1)*pas]))
listARbis.append(np.concatenate(listAR[n*pas:]))

arbis=np.concatenate(listARbis)

len(arbis)

dfbis=pd.DataFrame({'userId':arbis[:,0],'movieId':arbis[:,1],'itemId':arbis[:,2]})

dfbis.head()

dfbis.to_parquet('data/ratings.gzip',compression='gzip',index=False)
