==> AmazonBookReview/5 - result gestion.ipynb <==
import pandas as pd

def div(l1,a):
    l=list()
    for i in range(len(l1)):
        l.append(l1[i]/a)
    return l

def fusion(l1,l2):
    l=list()
    for i in range(len(l1)):
        l.append(l1[i]+l2[i])
    return l

def resultFinal(results):
    nbFold=4
    
    #l=list(results.values())
    n=list(results.keys())
    name=list()
    for i in range(0,len(n),nbFold):
        name.append(n[i][:-1])
    resultsFinal={}
    print (name)
    for i in range(len(n)):
        if i/nbFold-i//nbFold==0 :
            if i!=0 :
                resultsFinal[name[i//nbFold-1]]=div(lTemp,nbFold)
            lTemp=results[n[i]]
        else :
            lTemp=fusion(lTemp,results[n[i]])
        #print(lTemp[0])
    resultsFinal[name[-1]]=div(lTemp,nbFold)
    return resultsFinal

def inteName(i):
    if i==0:
        return "1-9"
    elif i<10:
        return str(int(i*10))+"-"+str(int(i*10+9))
    elif i>=50:
        return 'sup500'
    else :
        return str(int(i*10))+"-"+str(int(i*10+49))

df=pd.read_csv('data/resultRMSEbyInterval_1.csv')

df=df[:20]

df

nameColumms=list(df)

res={}
for index,row in df.iterrows():
    l=list()
    for i in nameColumms[1:]:
        l.append(row[i])
    res[row['algoName']]=l

resultsf=resultFinal(res)

resultsf

dfbis=pd.DataFrame(columns =nameColumms)
for key in resultsf:
    dfbis.loc[len(dfbis)]=[key]+resultsf[key]

dfbis.to_csv('data/resultRMSEbyInterval_V1.csv',index=False)

df2=dfbis.copy()

df2

dfbis=pd.concat([dfbis,df2])

dfbis
