==> AmazonBookReview/6 - RMSE by user.ipynb <==
import pandas as pd
import numpy as np
from math import sqrt

def inteName(i):
    if i==0:
        return "1-9"
    elif i<10:
        return str(int(i*10))+"-"+str(int(i*10+9))
    elif i>=50:
        return 'sup500'
    else :
        return str(int(i*10))+"-"+str(int(i*10+49))

dicoDf={}
for i in range(10) :
    dicoDf[i]=pd.read_csv('data/prediction_fold_'+str(i)+'.csv')

listDF=dicoDf.values()

dicoDf[0].head()

algoListe=["BaselineOnly", "SVD", "NMF", "SlopeOne", 
           "CoClustering", "NormalPredictor", "SVDpp",
           "KNNBasic", "KNNWithMeans", "KNNBaseline", "KNNWithZScore"]

nbAlgo=2

for df in listDF:
    for name in algoListe[:nbAlgo]:
        df["error "+name]=abs(df["rating"]-df["Predicted ratings "+name])

ncol=list(dicoDf[0])

errS=6

ncol

d={}
ar=df.to_numpy()

def rmse(l):
    s=0
    for i in l:
        s+=i*i
    return sqrt(s/len(l))

def avg(l):
    s=0
    for i in l:
        s+=i
    return s/len(l)

nameColumms=['UserId','Number of rated movies in the training dataset','Number of rated movies in the testing dataset']
for name in algoListe[:nbAlgo] :
    nameColumms.append('RMSE '+name+', per user')
for name in algoListe[:nbAlgo] :
    nameColumms.append('MAE '+name+', per user')

listDfRes=list()
for fold in range(10):
    print('--== ', fold,' ==--')
    d={}
    ar=dicoDf[fold].to_numpy()
    t=len(ar)
    for i in range(t) :
        ID=ar[i,0]
        if ID not in d :
            d[ID]={}
            for j in range(nbAlgo) :
                d[ID][j]=[ar[i,j+errS]]
        else :
            for j in range(nbAlgo) :
                d[ID][j].append(ar[i,j+errS])

    dfDico=pd.read_csv('dataCV3/user-nbRating_fold_'+str(fold)+'.csv')
    dRatingsTrain={}
    dRatingsTest={}
    for index, row in dfDico.iterrows():
        dRatingsTrain[row['userId']]=row['NBRatingTrain']
        dRatingsTest[row['userId']]=row['NBRatingTest']
    del dfDico

    d2={}
    for i in d:
        d2[i]={}
        for j in range(nbAlgo):
            d2[i][j]=rmse(d[i][j])

    d3={}
    for i in d:
        d3[i]={}
        for j in range(nbAlgo):
            d3[i][j]=avg(d[i][j])

    listAR=list()
    for i in d2 :
        listAR.append(np.array([[int(i),int(dRatingsTrain[i]), int(dRatingsTest[i]),
                                 d2[i][0], d2[i][1], 
                                 d3[i][0], d3[i][1]]]))
    arRes=np.concatenate(listAR)
    dfRes=pd.DataFrame({nameColumms[0]:arRes[:,0],nameColumms[1]:arRes[:,1],nameColumms[2]:arRes[:,2],nameColumms[3]:arRes[:,3],
                        nameColumms[4]:arRes[:,4],nameColumms[5]:arRes[:,5],nameColumms[6]:arRes[:,6]})

    dfRes.to_csv('data/uRMSE_uMAE_fold_'+str(fold)+'.csv',index=False)
    listDfRes.append(dfRes)

dfRes=pd.concat(listDfRes)
dfRes.to_csv('data/uRMSE_uMAE_full.csv',index=False)

dfRes.head()

del d,d3,d2,df

len(dfRes)

cpt=0
for df in listDF :
    print(len(df))
    cpt+=len(df)

    
print("tot = ",cpt)

cpt2=0
lres=list()
for fold in range(10):
    dfTemp=pd.read_csv('data/uRMSE_uMAE_fold_'+str(fold)+'.csv')
    print(len(dfTemp))
    cpt2+=len(dfTemp)
    lres.append(dfTemp)
    
print("tot = ",cpt2)

ar=dfRes.to_numpy()

t=len(ar)
d={}
for i in range(t) :
    ID=ar[i,1]
    if ID not in d:
        d[ID]={}
        for j in range(nbAlgo):
            d[ID][j]=[ar[i,3+j]]
    else :
        for j in range(nbAlgo):
            d[ID][j].append(ar[i,3+j])

d2={}
for i in d:
    d2[i]={}
    for j in range(nbAlgo):
        d2[i][j]=avg(d[i][j])

listAR=list()
for i in d2 :
    listAR.append(np.array([[int(i), d2[i][0], d2[i][1], d2[i][2]]]))

arRes=np.concatenate(listAR)

nameColumms=['Number of rated movies']
for name in algoListe[:nbAlgo] :
    nameColumms.append('avg RMSE '+name)

dfRes=pd.DataFrame({nameColumms[0]:arRes[:,0],nameColumms[1]:arRes[:,1],nameColumms[2]:arRes[:,2],nameColumms[3]:arRes[:,3]})
                    #nameColumms[4]:arRes[:,4],nameColumms[5]:arRes[:,5],nameColumms[6]:arRes[:,6]})

dfRes.head()

dfRes.to_csv('data/RMSEavg_cv.csv',index=False)
